refactor(lib): split bootstrap into env/db/services/window modules
Phase 1 Finding 6 from refactorRoadmap_2026-04-20.md. The `run()`
setup closure previously hid 11 initialization steps in one flat
block so any failure logged as `[startup]` with no per-concern
ownership. Splitting localizes each failure surface
(`[bootstrap/db]`, `[bootstrap/services]`, ...) and leaves `lib.rs`
holding only the `run()` entry point, `CancelRegistry`,
invoke_handler, and on_window_event.

- `bootstrap/env.rs` — `inherit_shell_path()`
- `bootstrap/db.rs` — DB path resolution + `db::init` + stale-state
  cleanup + `DbState` registration
- `bootstrap/services.rs` — HTTP API server, `PtyState`/`RawqIndexing`
  state, rawq daemon, bge-m3 embedder (sync + async fallback), orphan
  `sdk-url` process cleanup, vector backfill startup
- `bootstrap/window.rs` — window-state plugin logging + primary
  monitor centering fallback

`CancelRegistry` stays in `lib.rs`: it is cross-cutting state used
via `crate::CancelRegistry` from six command modules, and has no
natural sub-module owner. Out of scope for this Finding.

No user-visible behavior change. Baselines preserved:
- cargo test --lib: 295 passed
- cargo test --tests: 25 passed
